+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.1] / 2025-07-15
+### Features
+- Support `Navisworks` extensible data.
+### PackageBuilder
+- Add `NavisworksExtensibleData` to support `AppType` configuration.
+### Build
+- Update `build` files to use `LTS`.
+
 ## [2.0.0] / 2025-05-16 - 2025-06-05
 ### Features
 - Support `AutoCAD` bundle.
@@ -92,6 +100,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[2.0.1]: ../../compare/2.0.0...2.0.1
 [2.0.0]: ../../compare/1.0.6...1.2.0
 [1.0.6]: ../../compare/1.0.5...1.0.6
 [1.0.5]: ../../compare/1.0.4...1.0.5